begin refactor stdio_conn
