+++ internal/backend/rest_stdio_http2/stdio_conn/stdio_conn_other.go
@@ -1,6 +1,6 @@
 // +build !go1.10
 
-package rclone
+package stdio_conn
 
 import "time"
